+++ src/crave/frontend/RandvInterface.hpp
@@ -82,15 +82,15 @@ namespace crave {
     return *this;                            \
   }
 
-#define RANDV_PRIM_INTERFACE(Typename)             \
- public:                                           \
-  void gather_values(std::vector<long long>& ch) { \
-    ch.push_back(static_cast<long long>(value));   \
-  }                                                \
-  bool next() {                                    \
-    static distribution<Typename> dist;            \
-    value = dist.nextValue();                      \
-    return true;                                   \
+#define RANDV_PRIM_INTERFACE(Typename)           \
+ public:                                         \
+  void gather_values(std::vector<int64_t>& ch) { \
+    ch.push_back(static_cast<int64_t>(value));   \
+  }                                              \
+  bool next() {                                  \
+    static distribution<Typename> dist;          \
+    value = dist.nextValue();                    \
+    return true;                                 \
   }
 
 template <>
@@ -101,13 +101,13 @@ class randv<bool> : public randv_base<bool> {
 };
 
 // for all C/C++ built-in integer types
-#define RANDV_INTEGER_TYPE(typename)                       \
-  template <>                                              \
-  class randv<typename> : public randv_base<typename> {    \
-  RANDV_COMMON_INTERFACE(typename)                         \
-  RANDV_PRIM_INTERFACE(typename)                           \
-  RANDV_ARITHMETIC_INTERFACE(typename)                     \
-  RANDV_BITWISE_INTERFACE(typename)                        \
+#define RANDV_INTEGER_TYPE(typename)                    \
+  template <>                                           \
+  class randv<typename> : public randv_base<typename> { \
+    RANDV_COMMON_INTERFACE(typename)                    \
+    RANDV_PRIM_INTERFACE(typename)                      \
+    RANDV_ARITHMETIC_INTERFACE(typename)                \
+    RANDV_BITWISE_INTERFACE(typename)                   \
   };
 
 RANDV_INTEGER_TYPE(int)
