+++ src/crave/ir/UserConstraint.hpp
@@ -22,8 +22,6 @@
 #include "visitor/GetSupportSetVisitor.hpp"
 #include "visitor/ToDotNodeVisitor.hpp"
 
-
-
 namespace crave {
 
 int new_constraint_id();
@@ -48,12 +46,9 @@ struct UserConstraint {
   typedef NodePtr expression;
 
  protected:
-  UserConstraint(unsigned const id,
-                 expression const expr,
-                 std::string const& name,
-                 std::set<int>& support_vars,
-                 bool const soft = false,
-                 bool const cover = false,
+  UserConstraint(unsigned const id, expression const expr,
+                 std::string const& name, std::set<int>& support_vars,
+                 bool const soft = false, bool const cover = false,
                  bool const enabled = true)
       : id_(id),
         expr_(expr),
@@ -68,13 +63,10 @@ struct UserConstraint {
 
   template <typename ostream>
   friend ostream& operator<<(ostream& os, const UserConstraint& constr) {
-    os << constr.name_
-       << " is a "
-       << (constr.soft_ ? "soft" : "hard")
-       << " constraint and "
-       << (constr.enabled_ ? "enabled" : "disabled");
+    os << constr.name_ << " is a " << (constr.soft_ ? "soft" : "hard")
+       << " constraint and " << (constr.enabled_ ? "enabled" : "disabled");
     os << ", support vars =";
-    BOOST_FOREACH(int item, constr.support_vars_) { os << " " << item; }
+    BOOST_FOREACH (int item, constr.support_vars_) { os << " " << item; }
     return os;
   }
 
@@ -114,22 +106,12 @@ struct UserVectorConstraint : UserConstraint {
   friend struct ConstraintPartitioner;
 
  protected:
-  UserVectorConstraint(unsigned const id,
-                       expression const expr,
-                       std::string const& name,
-                       std::set<int>& support_vars,
-                       bool const unique,
-                       bool const soft = false,
-                       bool const cover = false,
-                       bool const enabled = true)
-      : UserConstraint(id,
-                       expr,
-                       name,
-                       support_vars,
-                       soft,
-                       cover,
-                       enabled),
-                       unique_(unique) {}
+  UserVectorConstraint(unsigned const id, expression const expr,
+                       std::string const& name, std::set<int>& support_vars,
+                       bool const unique, bool const soft = false,
+                       bool const cover = false, bool const enabled = true)
+      : UserConstraint(id, expr, name, support_vars, soft, cover, enabled),
+        unique_(unique) {}
 
  public:
   bool isUnique() { return unique_; }
@@ -170,13 +152,13 @@ struct ConstraintPartition {
   }
 
   bool containsVar(int id) const {
-      return support_vars_.find(id) != support_vars_.end();
+    return support_vars_.find(id) != support_vars_.end();
   }
 
   template <typename ostream>
   friend ostream& operator<<(ostream& os, const ConstraintPartition& cp) {
     os << "[ ";
-    BOOST_FOREACH(ConstraintPtr c, cp) { os << c->name() << " "; }
+    BOOST_FOREACH (ConstraintPtr c, cp) { os << c->name() << " "; }
     os << "]";
     os << std::flush;
     return os;
@@ -202,13 +184,11 @@ struct ConstraintManager {
 
   template <typename ostream>
   friend ostream& operator<<(ostream& os, const ConstraintManager& set) {
-    os << "Set "
-       << set.id_
-       << " has " << set.constraints_.size()
-       << " constraint(s) and has "
-       << (set.changed_ ? "" : "not ") << "changed" << std::endl;
-    BOOST_FOREACH(ConstraintPtr item, set.constraints_) {
-        os << item << std::endl;
+    os << "Set " << set.id_ << " has " << set.constraints_.size()
+       << " constraint(s) and has " << (set.changed_ ? "" : "not ") << "changed"
+       << std::endl;
+    BOOST_FOREACH (ConstraintPtr item, set.constraints_) {
+      os << item << std::endl;
     }
     os << std::flush;
     return os;
@@ -223,7 +203,8 @@ struct ConstraintManager {
         ite->second->enable();
         changed_ = true;
       } else {
-        LOG(INFO) << "  already enabled";}
+        LOG(INFO) << "  already enabled";
+      }
       return true;
     }
     LOG(INFO) << "not found";
@@ -239,7 +220,8 @@ struct ConstraintManager {
         ite->second->disable();
         changed_ = true;
       } else {
-        LOG(INFO) << "  already enabled";}
+        LOG(INFO) << "  already enabled";
+      }
       return true;
     }
     LOG(INFO) << "  not found";
@@ -256,20 +238,14 @@ struct ConstraintManager {
   void resetChanged() { changed_ = false; }
 
   template <typename Expr>
-  ConstraintPtr makeConstraint(std::string const& name,
-                               int c_id,
-                                Expr e,
-                               Context& ctx,
-                               bool const soft = false,
+  ConstraintPtr makeConstraint(std::string const& name, int c_id, Expr e,
+                               Context& ctx, bool const soft = false,
                                bool const cover = false) {
-    LOG(INFO) << "New "
-              << (soft ? "soft " : "")
-              << (cover ? "cover " : "")
-              << "constraint " << name << " in set "
-              << id_;
+    LOG(INFO) << "New " << (soft ? "soft " : "") << (cover ? "cover " : "")
+              << "constraint " << name << " in set " << id_;
 
     if (constr_map_.find(name) != constr_map_.end()) {
-        throw std::runtime_error("Constraint already exists.");
+      throw std::runtime_error("Constraint already exists.");
     }
 
     FixWidthVisitor fwv;
@@ -278,14 +254,21 @@ struct ConstraintManager {
     GetSupportSetVisitor gssv;
     n->visit(gssv);
 
-    ConstraintPtr c(boost::dynamic_pointer_cast<ForEach>(n) != 0
-                        ? new UserVectorConstraint(c_id, n, name, gssv.getSupportVars(), false, soft, cover)
-                        : (boost::dynamic_pointer_cast<Unique>(n) != 0
-                               ? new UserVectorConstraint(c_id, n, name, gssv.getSupportVars(), true, soft, cover)
-                               : new UserConstraint(c_id, n, name, gssv.getSupportVars(), soft, cover)));
+    ConstraintPtr c(
+        boost::dynamic_pointer_cast<ForEach>(n) != 0
+            ? new UserVectorConstraint(c_id, n, name, gssv.getSupportVars(),
+                                       false, soft, cover)
+            : (boost::dynamic_pointer_cast<Unique>(n) != 0
+                   ? new UserVectorConstraint(c_id, n, name,
+                                              gssv.getSupportVars(), true, soft,
+                                              cover)
+                   : new UserConstraint(c_id, n, name, gssv.getSupportVars(),
+                                        soft, cover)));
 
-    assert(!c->isSoft() || !c->isCover());              // soft cover constraint not defined/supported yet
-    assert(!c->isVectorConstraint() || !c->isCover());  // cover vector constraint not defined/supported yet
+    assert(!c->isSoft() ||
+           !c->isCover());  // soft cover constraint not defined/supported yet
+    assert(!c->isVectorConstraint() ||
+           !c->isCover());  // cover vector constraint not defined/supported yet
 
     changed_ = true;
     constraints_.push_back(c);
@@ -293,29 +276,34 @@ struct ConstraintManager {
     return constr_map_[name] = c;
   }
 
-  ConstraintPtr makeConstraint(std::string const& name, int c_id, NodePtr n, Context& ctx, bool const soft = false,
+  ConstraintPtr makeConstraint(std::string const& name, int c_id, NodePtr n,
+                               Context& ctx, bool const soft = false,
                                bool const cover = false) {
-    LOG(INFO) << "New "
-              << (soft ? "soft " : "")
-              << (cover ? "cover " : "")
-              << "constraint " << name << " in set "
-              << id_;
+    LOG(INFO) << "New " << (soft ? "soft " : "") << (cover ? "cover " : "")
+              << "constraint " << name << " in set " << id_;
 
     if (constr_map_.find(name) != constr_map_.end()) {
-        throw std::runtime_error(name+"Constraint already exists.");
+      throw std::runtime_error(name + "Constraint already exists.");
     }
 
     GetSupportSetVisitor gssv;
     n->visit(gssv);
 
-    ConstraintPtr c(boost::dynamic_pointer_cast<ForEach>(n) != 0
-                        ? new UserVectorConstraint(c_id, n, name, gssv.getSupportVars(), false, soft, cover)
-                        : (boost::dynamic_pointer_cast<Unique>(n) != 0
-                               ? new UserVectorConstraint(c_id, n, name, gssv.getSupportVars(), true, soft, cover)
-                               : new UserConstraint(c_id, n, name, gssv.getSupportVars(), soft, cover)));
+    ConstraintPtr c(
+        boost::dynamic_pointer_cast<ForEach>(n) != 0
+            ? new UserVectorConstraint(c_id, n, name, gssv.getSupportVars(),
+                                       false, soft, cover)
+            : (boost::dynamic_pointer_cast<Unique>(n) != 0
+                   ? new UserVectorConstraint(c_id, n, name,
+                                              gssv.getSupportVars(), true, soft,
+                                              cover)
+                   : new UserConstraint(c_id, n, name, gssv.getSupportVars(),
+                                        soft, cover)));
 
-    assert(!c->isSoft() || !c->isCover());              // soft cover constraint not defined/supported yet
-    assert(!c->isVectorConstraint() || !c->isCover());  // cover vector constraint not defined/supported yet
+    assert(!c->isSoft() ||
+           !c->isCover());  // soft cover constraint not defined/supported yet
+    assert(!c->isVectorConstraint() ||
+           !c->isCover());  // cover vector constraint not defined/supported yet
 
     changed_ = true;
     constraints_.push_back(c);
@@ -324,30 +312,27 @@ struct ConstraintManager {
   }
 
   template <typename Expr>
-  ConstraintPtr makeConstraint(std::string const& name,
-                               Expr e, Context& ctx,
+  ConstraintPtr makeConstraint(std::string const& name, Expr e, Context& ctx,
                                bool const soft = false,
                                bool const cover = false) {
     return makeConstraint(name, new_constraint_id(), e, ctx, soft, cover);
   }
 
   template <typename Expr>
-  ConstraintPtr makeConstraint(Expr e, Context& ctx,
-                               bool const soft = false,
+  ConstraintPtr makeConstraint(Expr e, Context& ctx, bool const soft = false,
                                bool const cover = false) {
     int id = new_constraint_id();
-    return makeConstraint("constraint_" +
-           boost::lexical_cast<std::string>(id), id, e, ctx, soft, cover);
+    return makeConstraint("constraint_" + boost::lexical_cast<std::string>(id),
+                          id, e, ctx, soft, cover);
   }
 
   std::ostream& printDotGraph(std::ostream& os) {
     ToDotVisitor visitor(os);
-    BOOST_FOREACH(ConstraintPtr c, constraints_) {
-      long a = reinterpret_cast<long>(&*c);
-      long b = reinterpret_cast<long>(&(*c->expr()));
-      os << "\t" << a << " [label=\""
-         << c->name() << (c->isSoft() ? " soft" : "")
-         << (c->isCover() ? " cover" : "")
+    BOOST_FOREACH (ConstraintPtr c, constraints_) {
+      int32_t a = reinterpret_cast<long>(&*c);
+      int32_t b = reinterpret_cast<long>(&(*c->expr()));
+      os << "\t" << a << " [label=\"" << c->name()
+         << (c->isSoft() ? " soft" : "") << (c->isCover() ? " cover" : "")
          << (!c->isEnabled() ? " disabled" : "") << "\"]" << std::endl;
       os << "\t" << a << " -> " << b << std::endl;
       c->expr()->visit(visitor);
@@ -375,13 +360,13 @@ struct ConstraintPartitioner {
 
   void mergeConstraints(const ConstraintManager& mng) {
     LOG(INFO) << "Merge set " << mng.id_ << " with set(s)";
-    BOOST_FOREACH(unsigned id, constr_mngs_) { LOG(INFO) << " " << id; }
+    BOOST_FOREACH (unsigned id, constr_mngs_) { LOG(INFO) << " " << id; }
     constr_mngs_.insert(mng.id_);
-    BOOST_FOREACH(ConstraintPtr c, mng.constraints_) {
+    BOOST_FOREACH (ConstraintPtr c, mng.constraints_) {
       if (c->isEnabled()) {
         if (c->isVectorConstraint())
           vec_constraints_.push_back(
-          boost::static_pointer_cast<UserVectorConstraint>(c));
+              boost::static_pointer_cast<UserVectorConstraint>(c));
 
         else
           constraints_.push_back(c);
@@ -396,30 +381,29 @@ struct ConstraintPartitioner {
       partitions_.push_back(cp);
     }
     LOG(INFO) << "Partition results of set(s)";
-    BOOST_FOREACH(unsigned id, constr_mngs_) { LOG(INFO) << " " << id; }
+    BOOST_FOREACH (unsigned id, constr_mngs_) { LOG(INFO) << " " << id; }
     LOG(INFO) << ": ";
 
     LOG(INFO) << "  " << vec_constraints_.size() << " vector constraint(s):";
-    BOOST_FOREACH(VectorConstraintPtr c, vec_constraints_) {
-        LOG(INFO) << "   " << c->name();
+    BOOST_FOREACH (VectorConstraintPtr c, vec_constraints_) {
+      LOG(INFO) << "   " << c->name();
     }
 
     LOG(INFO) << "  " << partitions_.size() << " constraint partition(s):";
     uint cnt = 0;
-    BOOST_FOREACH(ConstraintPartition & cp, partitions_) {
-        LOG(INFO) << "    #" << ++cnt << ": " << cp;
+    BOOST_FOREACH (ConstraintPartition& cp, partitions_) {
+      LOG(INFO) << "    #" << ++cnt << ": " << cp;
     }
   }
 
   std::vector<ConstraintPartition>& getPartitions() { return partitions_; }
 
   std::vector<VectorConstraintPtr>& getVectorConstraints() {
-      return vec_constraints_;
+    return vec_constraints_;
   }
 
  private:
-  void maximizePartition(ConstraintPartition& cp,
-                         ConstraintList& lc) {
+  void maximizePartition(ConstraintPartition& cp, ConstraintList& lc) {
     ConstraintPtr c = lc.front();
     lc.pop_front();
     cp.support_vars_ = c->support_vars_;
@@ -431,10 +415,8 @@ struct ConstraintPartitioner {
       while (ite != lc.end()) {
         c = *ite;
         std::vector<int> v_intersection;
-        std::set_intersection(cp.support_vars_.begin(),
-                              cp.support_vars_.end(),
-                              c->support_vars_.begin(),
-                              c->support_vars_.end(),
+        std::set_intersection(cp.support_vars_.begin(), cp.support_vars_.end(),
+                              c->support_vars_.begin(), c->support_vars_.end(),
                               std::back_inserter(v_intersection));
         if (!v_intersection.empty()) {
           changed = true;
